requirements.txt
docx~=0.2.4
paramiko~=2.8.0
PyMySQL~=1.0.2
redis~=4.2.2
SQLAlchemy~=1.4.36
requests~=2.26.0
simplejson~=3.17.6
jsonpath~=0.82
PyYAML~=6.0
pandas~=1.3.5
pytest~=6.2.5
Faker~=13.4.0
DingtalkChatbot~=1.5.3
colorlog~=6.6.0